+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===
 - Use file cache on load error - GitHub issue #20
 - Error converting value to number - GitHub issue #21
+- Problem with punycode - GitHub issue #22
 
 0.3.11
 ===